jouster-llm-knowledge-extractor/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.9.1
SQLAlchemy==2.0.34
python-dotenv==1.0.1
nltk==3.9.1
openai==1.50.2
httpx==0.27.2
pytest==8.3.3
requests==2.32.3
